Extract BuildConfig field from Onliner`s URL closes #93
