+++ app/src/main/java/kt/school/starlord/model/httpurlconnection/PageLoader.kt
@@ -1,10 +1,11 @@
-package kt.school.starlord.model
+package kt.school.starlord.model.httpurlconnection
 
+import kt.school.starlord.BuildConfig
 import java.net.HttpURLConnection
 import java.net.URL
 
 class PageLoader {
-    private val url: URL = URL("https://baraholka.onliner.by")
+    private val url: URL = URL(BuildConfig.BARAHOLKA_ONLINER_URL)
 
     fun getPage(): String {
         val urlConnection: HttpURLConnection = url.openConnection() as HttpURLConnection
