+++ gradle.properties
@@ -2,3 +2,6 @@ org.gradle.jvmargs=-Xmx1536m
 android.useAndroidX=true
 android.enableJetifier=true
 kotlin.code.style=official
+
+# App constants
+baraholkaOnlinerUrl = "https://baraholka.onliner.by"
